requirements.txt
pipreqs
boto3==1.17.5
requests
pyyaml